Settings and S2 control added
Added a few settings.

Added a way to control the S2 button to trigger a flow.
Since the Dimmer 2 uses scene_activation, that isn’t supported yet. I
tried to find a work around.

I added association group 4, this sends an on/off command to the homey
when S2 is pressed. Since the S1 uses multi_level this shouldn’t be an
issue.
When a BASIC report is being received, it triggers a flow trigger card.