fix 2 issues
"value is out of bounds":
was over doing it with the (input) setting(s),
and doing a change on input(s) that is/was not needed,
or rather it is causing unwanted behaviour. (add 8 to all inputs if an
analog input is attached, not needed with the separate inputs)
not 100% sure this will fix the issue,
but can't think of another reason why the value is getting too high,
and couldn't reproduce the crash.

"Cannot read property 'w' of null":
accidentily used the wrong cache of a (the) device

also noticed a typo in transition mode 2 settings

and fixed an (unnoticed?) issue when trying to setSettings on change of
strip type and the input value is +8, the setSettings isn't even needed
at all anymore (was needed in a previous itteration of the driver), so
removed it completely.