FGRGBW-441: custom mobile card. FGSD-002: textual edit
Added a custom mobile card for the fgrgbw, since in 1.3 the
measure_power and meter_power will show in the mobile card
automatically.
The sensors will make the dim slider overlap the color selector. (on
android mobile worse then the mobile card in chrome)
Will create a issue about the weird overlapping issue, but this is best
option for controlling the rgbw controller anyway.

There was a textual mistake in the hint of the fgsd-002, this will fix
it.